break out upstream axp192 code
